@DifferentialOrange 2021-12-20
PR is ready for review, but since it depends on #244 improvements, I will not request any reviews for now (until #244 is merged).